requirement.txt
selenium
youtube_dl
py-cord
python-dotenv